refactor: standardize modules names to improve public API surface
