+++ docs/api/client.md
@@ -1,5 +1,5 @@
 # cb_events.client
 
-::: cb_events.client
+::: cb_events._client
 
 <sub>This section is generated directly from package docstrings.</sub>
