+++ docs/api/config.md
@@ -1,5 +1,5 @@
 # cb_events.config
 
-::: cb_events.config
+::: cb_events._config
 
 <sub>This section is generated directly from package docstrings.</sub>
