+++ docs/api/exceptions.md
@@ -1,5 +1,5 @@
 # cb_events.exceptions
 
-::: cb_events.exceptions
+::: cb_events._exceptions
 
 <sub>This section is generated directly from package docstrings.</sub>
