+++ docs/api/models.md
@@ -1,5 +1,5 @@
 # cb_events.models
 
-::: cb_events.models
+::: cb_events._models
 
 <sub>This section is generated directly from package docstrings.</sub>
