+++ docs/api/router.md
@@ -1,5 +1,5 @@
 # cb_events.router
 
-::: cb_events.router
+::: cb_events._router
 
 <sub>This section is generated directly from package docstrings.</sub>
