+++ docs/configuration.md
@@ -108,7 +108,7 @@ logging.getLogger("cb_events").setLevel(logging.DEBUG)
 **Example DEBUG output**:
 
 ```text
-DEBUG:cb_events.client:Polling https://eventsapi.chaturbate.com/events/user/************************/?timeout=10
-DEBUG:cb_events.client:Received 1 events for user
-DEBUG:cb_events.router:Dispatching chatMessage event 1775683684418-0 to 2 handlers
+DEBUG:cb_events._client:Polling https://eventsapi.chaturbate.com/events/user/************************/?timeout=10
+DEBUG:cb_events._client:Received 1 events for user
+DEBUG:cb_events._router:Dispatching chatMessage event 1775683684418-0 to 2 handlers
 ```
