+++ src/cb_events/_exceptions.py
