+++ src/cb_events/_models.py
